[Add]receiptscontroller追加
