+++ app/helpers/receipts_helper.rb
@@ -1,0 +1,2 @@
+module ReceiptsHelper
+end
